provision: Cleanup pgsql install script

+++ provision/scripts/install_pgsql.sh
@@ -19,12 +19,12 @@ yum_install postgresql93-{server,contrib,devel,plpython} postgis2_93
 # SKP compatibility
 install -d -o postgres -g postgres -m 700 "/data/pg_data"
 install -d -o postgres -g postgres -m 700 "/data/pg_log"
-if [[ ! -d /app ]]; then
-	ln -s /srv /app
+ln -s /data/pg_data "$PGSQL_BASE/data"
+ln -s /data/pg_log "$PGSQL_BASE/logs"
+[[ -L /app ]] || ln -s /srv /app
+if [[ ! -d $PGSQL_BASE/bin ]]; then
 	install -d -o root -g root "$PGSQL_BASE/bin"
 	ln -s /usr/pgsql-9.3/bin/* "$PGSQL_BASE/bin"
-	ln -s /data/pg_data "$PGSQL_BASE/data"
-	ln -s /data/pg_log "$PGSQL_BASE/logs"
 fi
 
 echo "export PATH=$PATH:/usr/pgsql-9.3/bin" > /etc/profile.d/pgsql.sh
